requirements.txt
kivy
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
pytz
google-generativeai
